# Speed up shellcraft loading and clean up code a bit