Old:

```
$ time shellcraft
...
shellcraft  0.28s user 0.17s system 98% cpu 0.455 total
```

New:

```
$ time shellcraft
...
shellcraft  0.17s user 0.10s system 102% cpu 0.272 total
```