@zachriggle 2015-01-26
Ah, I thought it's supposed to be for a C/Python/Perl-compatible string.  Will fix.