@IdolfHatler 2015-01-26
Well I guess that too. I've just never used it for that. :)