@zachriggle 2015-02-25
Nope, if forgotten about it. I just need to put the old string print code back in.