@zachriggle 2015-02-25
I've rebased this against the current `Gallopsled/pwntools/master` and removed the changes to string-printing per @Idolf's comments